Upgrade from ASM6 to ASM8
These two APIs seem to be completely compatible, or at minimum they
don't seem to loose any gadgets on tests that I've done. That being
said, the only benefit of doing this (it seems) is to be able to
move from using Java 8 as the compiler to Java 11 (maybe 14, haven't
tested that one yet). So this patch is pretty optional.